another minor refactor
