+++ lib/pages/dashboard.dart
@@ -1,6 +1,6 @@
 import 'package:expense_manager/components/dashboard_widgets/dashboard_widget.dart';
-import 'package:expense_manager/components/dashboard_widgets/expense_pie_chart/expense_pie_chart.dart';
-import 'package:expense_manager/components/dashboard_widgets/expense_vs_income_pie_chart/expense_vs_income_pie_chart.dart';
+import 'package:expense_manager/components/dashboard_widgets/expense_pie_chart.dart';
+import 'package:expense_manager/components/dashboard_widgets/expense_vs_income_pie_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
